+++ dzVents/documentation/history.md
@@ -3,7 +3,10 @@ __NOTOC__
 ```
 ## [3.1.10] ##
 - Added LOG_STATUS, adds a status log message into the system
+- Added 'Logging' example
 - Log levels are now handled internally as normal levels, meaning for example that error logs will be logged as errors, debug logs will be handles as debug logs (you have to enable debug logs as startup parameter)
+- Removed LOG_MODULE_EXEC_INFO from documentation as it should only be used internally (start/stop script notifications)
+- Removed unused log level 'LOG_WARNING'
 
 ## [3.1.9] ##
 - Added historical data helper <code>med(fromIndex, toIndex, default)</code> to calculate the median